requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Configuration
pyyaml>=6.0.0
configparser>=6.0.0

# Logging and utilities
colorlog>=6.7.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
responses>=0.23.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0